# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Australian KPIs API",
    "description": "Unified API for Australian public economic data.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/dataflows": {
      "get": {
        "tags": ["dataflows"],
        "summary": "`GET /v1/dataflows`.",
        "operationId": "listDataflows",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "description": "Optional source id filter, e.g. abs.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1
            }
          },
          {
            "name": "frequency",
            "in": "query",
            "description": "Optional publication frequency filter.",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "pattern": "^(annual|quarterly|monthly|weekly|daily|irregular)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dataflow catalog page.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Public CDN cache policy."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataflowsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/dataflows/{id}": {
      "get": {
        "tags": ["dataflows"],
        "summary": "`GET /v1/dataflows/{id}`.",
        "operationId": "getDataflow",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Dataflow id.",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dataflow metadata and dimensions.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Public CDN cache policy."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataflowDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path parameter.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Dataflow not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/dataflows/{id}/codelists/{dim}": {
      "get": {
        "tags": ["dataflows"],
        "summary": "`GET /v1/dataflows/{id}/codelists/{dim}`.",
        "operationId": "getDataflowCodelist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Dataflow id.",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1
            }
          },
          {
            "name": "dim",
            "in": "path",
            "description": "Dimension id.",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Codelist for the requested dimension.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Public CDN cache policy."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataflowCodelistResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path parameter.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Dataflow or dimension not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/observations": {
      "get": {
        "tags": ["observations"],
        "summary": "`GET /v1/observations`.",
        "operationId": "listObservations",
        "parameters": [
          {
            "name": "dataflow",
            "in": "query",
            "description": "Required dataflow id, e.g. abs.cpi.",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1
            }
          },
          {
            "name": "dimensions[]",
            "in": "query",
            "description": "Dimension filters as repeated key=value values. The handler also accepts dimensions[region]=AUS.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "Inclusive lower time bound as YYYY-MM-DD or RFC3339.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "minLength": 1
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "Inclusive upper time bound as YYYY-MM-DD or RFC3339.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "minLength": 1
            }
          },
          {
            "name": "frequency",
            "in": "query",
            "description": "Optional dataflow frequency filter, or weekly/monthly/quarterly rollup grain.",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "pattern": "^(annual|quarterly|monthly|weekly|daily|irregular)$"
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Response format: json, csv, or parquet.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 7,
              "minLength": 3,
              "pattern": "^(json|csv|parquet)$"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor from the previous page.",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size. JSON/CSV are capped at 10000 rows; Parquet bulk exports are capped at 1000000 rows.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1000000,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Observation page.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Public CDN cache policy."
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Weak entity tag for this page."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationsResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              },
              "application/vnd.apache.parquet": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "The client's cached page is still fresh.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Public CDN cache policy."
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Weak entity tag for this page."
              }
            }
          },
          "400": {
            "description": "Invalid query string.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Dataflow not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Bulk stream admission capacity is exhausted.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Short-request admission or a durable dependency is unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "Request timed out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/openapi.json": {
      "get": {
        "tags": [],
        "summary": "`GET /v1/openapi.json`.",
        "operationId": "openapi",
        "responses": {
          "200": {
            "description": "Current OpenAPI document.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "OpenAPI generation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "Request timed out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scorecards/aps/config": {
      "get": {
        "tags": ["scorecards"],
        "summary": "`GET /v1/scorecards/aps/config`.",
        "operationId": "getApsScorecardConfig",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "description": "Config version; omitted selects the current version.",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Versioned APS scorecard config.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Public CDN cache policy."
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Strong entity tag for the config JSON."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardConfig"
                }
              }
            }
          },
          "304": {
            "description": "The client's cached config is still fresh.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Public CDN cache policy."
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Strong entity tag for the config JSON."
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scorecards/aps/history": {
      "get": {
        "tags": ["scorecards"],
        "summary": "`GET /v1/scorecards/aps/history`.",
        "operationId": "listApsScorecardHistory",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "Inclusive lower snapshot date in `YYYY-MM-DD` form.",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "Inclusive upper snapshot date in `YYYY-MM-DD` form.",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          },
          {
            "name": "view",
            "in": "query",
            "description": "Revision view, defaulting to original as-published values.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/HistoryView"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum points returned, from 1 through 1,000.",
            "required": false,
            "schema": {
              "type": ["integer", "null"],
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Time-ordered APS scorecard snapshots.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Public CDN cache policy."
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Strong entity tag for the history JSON."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApsSnapshotSummary"
                  }
                }
              }
            }
          },
          "304": {
            "description": "The client's cached history response is still fresh.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Public CDN cache policy."
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Strong entity tag for the history JSON."
              }
            }
          },
          "400": {
            "description": "Invalid query string.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scorecards/aps/latest": {
      "get": {
        "tags": ["scorecards"],
        "summary": "`GET /v1/scorecards/aps/latest`.",
        "operationId": "getApsScorecardLatest",
        "responses": {
          "200": {
            "description": "Latest APS scorecard snapshot.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Public CDN cache policy."
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Strong entity tag for the latest snapshot JSON."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishedApsSnapshot"
                }
              }
            }
          },
          "304": {
            "description": "The client's cached latest snapshot is still fresh.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Public CDN cache policy."
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Strong entity tag for the latest snapshot JSON."
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scorecards/aps/snapshots/{id}": {
      "get": {
        "tags": ["scorecards"],
        "summary": "`GET /v1/scorecards/aps/snapshots/{id}`.",
        "operationId": "getApsScorecardSnapshot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Snapshot revision UUID.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Immutable APS snapshot revision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishedApsSnapshot"
                }
              }
            }
          },
          "404": {
            "description": "Snapshot not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/search": {
      "get": {
        "tags": ["search"],
        "summary": "`GET /v1/search`.",
        "operationId": "searchCatalog",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search text.",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results, capped at 100.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 65535,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked catalog search results.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Public CDN cache policy."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/series/{dataflow}/{series_key}": {
      "get": {
        "tags": ["series"],
        "summary": "`GET /v1/series/{dataflow}/{series_key}`.",
        "operationId": "getSeries",
        "parameters": [
          {
            "name": "dataflow",
            "in": "path",
            "description": "Dataflow id.",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1
            }
          },
          {
            "name": "series_key",
            "in": "path",
            "description": "64-character series key hex digest.",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "minLength": 64,
              "pattern": "^[0-9a-f]{64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Series metadata and latest observation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesLookupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path parameter.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Series not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataflowsResponse": {
        "type": "object",
        "description": "Response envelope for `GET /v1/dataflows`.",
        "required": ["dataflows"],
        "properties": {
          "dataflows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Dataflow"
            },
            "description": "Matching dataflows, ordered by source then dataflow id."
          }
        }
      },
...
      "HistoryView": {
        "type": "string",
        "description": "Revision view used by APS history reads.",
        "enum": ["as_published", "latest"]
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 problem details body.",
        "required": ["type", "title", "status"],
        "properties": {
          "detail": {
            "type": ["string", "null"],
            "description": "Request-specific detail, if any."
          },
          "instance": {
            "type": ["string", "null"],
            "description": "Resource-specific identifier, if any."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code for this problem.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary."
          },
          "type": {
            "type": "string",
            "description": "Problem type URI."
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}